[Oztechan/CCC#2602] Move SettingsView to SettingsRootView
