+++ ios/CCC/UI/Settings/SettingsView.swift
@@ -14,33 +14,18 @@ import Res
 import SwiftUI
 
 struct SettingsView: View {
-    @StateObject var observable = ObservableSEEDViewModel<
-        SettingsState,
-        SettingsEffect,
-        SettingsEvent,
-        SettingsData,
-        SettingsViewModel
-    >()
     @Environment(\.colorScheme) var colorScheme
-    @EnvironmentObject private var navigationStack: NavigationStackCompat
-    @State var premiumViewVisibility = false
-    @State var emailViewVisibility = false
-    @State var webViewVisibility = false
-    @State var isAdsAlreadyDisabledSnackShown = false
-    @State var isAlreadySyncedSnackShown = false
-    @State var isSynchronisingShown = false
-    @State var isSyncedSnackShown = false
 
-    private let analyticsManager: AnalyticsManager = koin.get()
-
-    var onBaseChange: ((String) -> Void)
+    var event: SettingsEvent
+    var state: SettingsState
+    var shouldShowBannerAd: Bool
 
     var body: some View {
         ZStack {
             Res.colors().background_strong.get().edgesIgnoringSafeArea(.all)
 
             VStack {
-                SettingsToolbarView(backEvent: observable.event.onBackClick)
+                SettingsToolbarView(backEvent: event.onBackClick)
 
                 Form {
                     SettingsItemView(
@@ -48,9 +33,9 @@ struct SettingsView: View {
                         title: Res.strings().settings_item_currencies_title.get(),
                         subTitle: Res.strings().settings_item_currencies_sub_title.get(),
                         value: Res.strings().settings_active_item_value.get(
-                            parameter: observable.state.activeCurrencyCount
+                            parameter: state.activeCurrencyCount
                         ),
-                        onClick: observable.event.onCurrenciesClick
+                        onClick: event.onCurrenciesClick
                     )
 
                     SettingsItemView(
@@ -58,17 +43,17 @@ struct SettingsView: View {
                         title: Res.strings().settings_item_watchers_title.get(),
                         subTitle: Res.strings().settings_item_watchers_sub_title.get(),
                         value: Res.strings().settings_active_item_value.get(
-                            parameter: observable.state.activeWatcherCount
+                            parameter: state.activeWatcherCount
                         ),
-                        onClick: observable.event.onWatchersClick
+                        onClick: event.onWatchersClick
                     )
 
                     SettingsItemView(
                         imgName: "crown.fill",
                         title: Res.strings().settings_item_premium_title.get(),
                         subTitle: Res.strings().settings_item_premium_sub_title_no_ads.get(),
-                        value: getPremiumText(premiumStatus: observable.state.premiumStatus),
-                        onClick: observable.event.onPremiumClick
+                        value: getPremiumText(premiumStatus: state.premiumStatus),
+                        onClick: event.onPremiumClick
                     )
 
                     SettingsItemView(
@@ -76,7 +61,7 @@ struct SettingsView: View {
                         title: Res.strings().settings_item_sync_title.get(),
                         subTitle: Res.strings().settings_item_sync_sub_title.get(),
                         value: "",
-                        onClick: observable.event.onSyncClick
+                        onClick: event.onSyncClick
                     )
 
                     if MailView.canSendEmail() {
@@ -85,7 +70,7 @@ struct SettingsView: View {
                             title: Res.strings().settings_item_feedback_title.get(),
                             subTitle: Res.strings().settings_item_feedback_sub_title.get(),
                             value: "",
-                            onClick: observable.event.onFeedBackClick
+                            onClick: event.onFeedBackClick
                         )
                     }
 
@@ -94,96 +79,24 @@ struct SettingsView: View {
                         title: Res.strings().settings_item_on_github_title.get(),
                         subTitle: Res.strings().settings_item_on_github_sub_title.get(),
                         value: "",
-                        onClick: observable.event.onOnGitHubClick
+                        onClick: event.onOnGitHubClick
                     )
 
                     SettingsItemView(
                         imgName: "textformat.123",
                         title: Res.strings().settings_item_version_title.get(),
                         subTitle: Res.strings().settings_item_version_sub_title.get(),
-                        value: observable.state.version,
+                        value: state.version,
                         onClick: {}
                     )
                 }.edgesIgnoringSafeArea(.bottom)
                     .withClearBackground(color: Res.colors().background.get())
 
-                if observable.viewModel.shouldShowBannerAd() {
+                if shouldShowBannerAd {
                     AdaptiveBannerAdView(unitID: "BANNER_AD_UNIT_ID_SETTINGS").adapt()
                 }
             }
             .navigationBarHidden(true)
-        }
-        .popup(
-            isPresented: $isAdsAlreadyDisabledSnackShown,
-            type: .toast,
-            autohideIn: 2.0
-        ) {
-            SnackView(text: Res.strings().txt_you_already_have_premium.get())
-        }
-        .popup(
-            isPresented: $isAlreadySyncedSnackShown,
-            type: .toast,
-            autohideIn: 2.0
-        ) {
-            SnackView(text: Res.strings().txt_already_synced.get())
-        }
-        .popup(
-            isPresented: $isSynchronisingShown,
-            type: .toast,
-            autohideIn: 2.0
-        ) {
-            SnackView(text: Res.strings().txt_synchronising.get())
-        }
-        .popup(
-            isPresented: $isSyncedSnackShown,
-            type: .toast,
-            autohideIn: 2.0
-        ) {
-            SnackView(text: Res.strings().txt_synced.get())
-        }
-        .sheet(isPresented: $premiumViewVisibility) {
-            PremiumView(premiumViewVisibility: $premiumViewVisibility)
-        }
-        .sheet(isPresented: $emailViewVisibility) {
-            MailView(isShowing: $emailViewVisibility)
-        }
-        .sheet(isPresented: $webViewVisibility) {
-            WebView(url: NSURL(string: Res.strings().github_url.get())! as URL)
-        }
-        .onAppear {
-            observable.startObserving()
-            analyticsManager.trackScreen(screenName: ScreenName.Settings())
-        }
-        .onDisappear { observable.stopObserving() }
-        .onReceive(observable.effect) { onEffect(effect: $0) }
-    }
-
-    // swiftlint:disable:next cyclomatic_complexity
-    private func onEffect(effect: SettingsEffect) {
-        logger.i(message: { "SettingsView onEffect \(effect.description)" })
-        switch effect {
-        case is SettingsEffect.Back:
-            navigationStack.pop()
-        case is SettingsEffect.OpenCurrencies:
-            navigationStack.push(CurrenciesRootView(onBaseChange: onBaseChange))
-        case is SettingsEffect.OpenWatchers:
-            navigationStack.push(WatchersView())
-        case is SettingsEffect.FeedBack:
-            emailViewVisibility.toggle()
-        case is SettingsEffect.OnGitHub:
-            webViewVisibility.toggle()
-        case is SettingsEffect.Synchronising:
-            isSynchronisingShown.toggle()
-        case is SettingsEffect.Synchronised:
-            isSyncedSnackShown.toggle()
-        case is SettingsEffect.OnlyOneTimeSync:
-            isAlreadySyncedSnackShown.toggle()
-        case is SettingsEffect.AlreadyPremium:
-            isAdsAlreadyDisabledSnackShown.toggle()
-        case is SettingsEffect.Premium:
-            premiumViewVisibility.toggle()
-        default:
-            logger.i(message: { "SettingsView unknown effect" })
         }
     }
